requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
chromadb
sentence-transformers
pypdf
python-docx